% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_first_birth}
\alias{age_first_birth}
\title{Mean age at first birth 1970-2011}
\format{
A data frame with 588 rows and 5 variables:
\describe{
  \item{\code{Country}}{character. Country name.}
  \item{\code{iso3}}{character. ISO Alpha 3 code.}
  \item{\code{period}}{character. Earlier, Middle, Latest}
  \item{\code{year_original}}{character. Year of survey. Includes some ranges, e.g., 1975-1976.}
  \item{\code{age_first_birth}}{double. Mean age at first birth.}
  \item{\code{year}}{integer. Year of survey. Year ranges recoded to first or last year.}
}
}
\source{
World Fertility Report 2012
}
\usage{
age_first_birth
}
\description{
Mean age at first birth
}
\keyword{datasets}
