% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asfr}
\alias{asfr}
\title{Age-Specific Fertility Rates 1950-2015}
\format{
A data frame with 2600 rows and 11 variables:
\describe{
  \item{\code{country}}{character. Country name.}
  \item{\code{iso_code}}{double. Country code}
  \item{\code{year}}{character. Year (in 5-year blocks from 1950-1955 to 2010-2015).}
  \item{\code{15-19}}{double. 15-19 year olds.}
  \item{\code{20-24}}{double. 20-24 year olds.}
  \item{\code{25-29}}{double. 25-29 year olds.}
  \item{\code{30-34}}{double. 30-34 year olds.}
  \item{\code{35-39}}{double. 35-39 year olds.}
  \item{\code{40-44}}{double. 40-44 year olds.}
  \item{\code{45-49}}{double. 45-50 year olds.}
  \item{\code{iso3}}{character. ISO Alpha 3 code.}
}
}
\source{
\url{http://esa.un.org/unpd/wpp/Download/Standard/Fertility/}
}
\usage{
asfr
}
\description{
Age-specific fertility rates computed by the UN Department of Economic and Social
Affairs, Population Division. World Population Prospects. The 2015 Revision.
\url{http://esa.un.org/unpd/wpp/publications/files/key_findings_wpp_2015.pdf}
}
\keyword{datasets}
