% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastfeeding}
\alias{breastfeeding}
\title{Breastfeeding rates 2010-2015}
\format{
A data frame with 197 rows and 12 variables:
\describe{
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{country}}{character. Country name.}
  \item{\code{year_initial}}{character. Reference year for data on breastfeeding initiation.}
  \item{\code{percent_initial}}{double. Percent breastfeeding initiation.}
  \item{\code{year_6_months}}{character. Reference year for data on exclusive breastfeeding at 6 months.}
  \item{\code{percent_6_months}}{double. Percent exclusive breastfeeding at 6 months.}
  \item{\code{year_food_intro}}{character. Reference year for data on introduction of solid food at 6-8 months.}
  \item{\code{percent_food_intro}}{double. Percent introduction of solid food at 6-8 months.}
  \item{\code{year_1_year}}{character. Reference year for data on exclusive breastfeeding at 1 year.}
  \item{\code{percent_1_year}}{double. Percent exclusive breastfeeding at 1 year}
  \item{\code{year_2_year}}{character. Reference year for data on exclusive breastfeeding at 2 years.}
  \item{\code{percent_2_year}}{double. Percent exclusive breastfeeding at 2 years}
  \item{\code{year_min_accept}}{character. Reference year for data on minimum acceptable diet 6-23 months.}
  \item{\code{percent_min_accept}}{double. Percent minimum acceptable diet 6-23 months.}
  \item{\code{year_min_diverse}}{character. Reference year for data on minimum diet diversity 6-23 months.}
  \item{\code{percent_min_diverse}}{double. Percent minimum diet diversity 6-23 months.}
  \item{\code{year_min_freq}}{character. Reference year for data on minimum meal frequency 6-23 months.}
  \item{\code{percent_min_freq}}{double. Percent minimum meal frequency 6-23 months.}
}
}
\source{
\url{http://data.unicef.org/nutrition/iycf.html}
}
\usage{
breastfeeding
}
\description{
\itemize{
\item Early initiation of breastfeeding (<1hour, \%): Percentage of infants who are put to the breast within one hour of birth.
\item Exclusive breastfeeding (<6 months, \%): Percentage of children aged 0–5 months who are fed exclusively with breast milk in the 24 hours prior to the survey.
\item Introduction to solid, semi-solid or soft foods (6-8 months, \%): Percentage of children aged 6–8 months who received solid, semi-solid or soft foods in the 24 hours prior to the survey.
\item Breastfeeding at 1 year (12-15 months, \%): Percentage of children aged 12–15 months who received breast milk in the 24 hours prior to the survey.
\item Breastfeeding at age 2 (20-23 months, \%): Percentage of children aged 20–23 months who received breast milk in the 24 hours prior to the survey.
\item Minimum Diet Diversity (6-23 months, \%): Percentage of children 6–23 months of age who had ate from at least 4 (out of 7 pre-defined food groups) during the previous day.
\item Minimum Acceptable Diet (6-23 months, \%): Percentage of breastfed children 6–23 months of age who had at least the minimum dietary diversity and the minimum meal frequency during the previous day AND percentage of non-breastfed children 6–23 months of age who received at least 2 milk feedings and had at least the minimum dietary diversity not including milk feeds and the minimum meal frequency during the previous day.
}
}
\keyword{datasets}
