% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ciri}
\alias{ciri}
\title{CIRI Gender inequality indices 1981-2011}
\format{
A data frame with 5983 rows and 4 variables:
\describe{
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{YEAR}}{integer. Year.}
  \item{\code{WECON}}{integer. Women's Economic Rights}
  \item{\code{WOPOL}}{integer. Women's Political Rights}
}
}
\source{
Cingranelli, David L., David L. Richards, and K. Chad Clay. 2014. "The CIRI Human Rights Dataset."  \url{http://www.humanrightsdata.com}. Version 2014.04.14

\url{http://www.humanrightsdata.com/p/data-documentation.html}
}
\usage{
ciri
}
\description{
Gender inequality indices from the Cingranelli-Richards (CIRI) data-set.
}
\details{
WECON: A score of 0 indicates that there were no economic rights
for women in law and that systematic discrimination based on sex
may have been built into law. A score of 1 indicates that women
had some economic rights under law, but these rights were not
effectively enforced. A score of 2 indicates that women had some
economic rights under law, and the government effectively enforced
these rights in practice while still allowing a low level of
discrimination against women in economic matters. Finally, a score
of 3 indicates that all or nearly all of women’s economic rights
were guaranteed by law and the government fully and vigorously
enforces these laws in practice.

WOPOL: A score of 0 indicates that women’s political rights were
not guaranteed by law during a given year. A score of 1 indicates
that women’s political rights were guaranteed in law, but severely
prohibited in practice. A score of 2 indicates that women’s political
rights were guaranteed in law, but were still moderately prohibited
in practice. Finally, a score of 3 indicates that women’s political
rights were guaranteed in both law and practice.
}
\keyword{datasets}
