% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contraception}
\alias{contraception}
\title{Contraceptive prevalence 1970-2030}
\format{
A data frame with 585 rows and 64 variables:
\describe{
  \item{\code{country}}{character. Country name}
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{type}}{character. Modern, Traditional, or Any.}
  \item{\code{bc1970}}{double. Percent using contraception.}
  \item{\code{bc1971}}{double. Percent using contraception}
  \item{\code{bc1972}}{double. Percent using contraception}
  \item{\code{bc1973}}{double. Percent using contraception}
  \item{\code{bc1974}}{double. Percent using contraception}
  \item{\code{bc1975}}{double. Percent using contraception}
  \item{\code{bc1976}}{double. Percent using contraception}
  \item{\code{bc1977}}{double. Percent using contraception}
  \item{\code{bc1978}}{double. Percent using contraception}
  \item{\code{bc1979}}{double. Percent using contraception}
  \item{\code{bc1980}}{double. Percent using contraception}
  \item{\code{bc1981}}{double. Percent using contraception}
  \item{\code{bc1982}}{double. Percent using contraception}
  \item{\code{bc1983}}{double. Percent using contraception}
  \item{\code{bc1984}}{double. Percent using contraception}
  \item{\code{bc1985}}{double. Percent using contraception}
  \item{\code{bc1986}}{double. Percent using contraception}
  \item{\code{bc1987}}{double. Percent using contraception}
  \item{\code{bc1988}}{double. Percent using contraception}
  \item{\code{bc1989}}{double. Percent using contraception}
  \item{\code{bc1990}}{double. Percent using contraception}
  \item{\code{bc1991}}{double. Percent using contraception}
  \item{\code{bc1992}}{double. Percent using contraception}
  \item{\code{bc1993}}{double. Percent using contraception}
  \item{\code{bc1994}}{double. Percent using contraception}
  \item{\code{bc1995}}{double. Percent using contraception}
  \item{\code{bc1996}}{double. Percent using contraception}
  \item{\code{bc1997}}{double. Percent using contraception}
  \item{\code{bc1998}}{double. Percent using contraception}
  \item{\code{bc1999}}{double. Percent using contraception}
  \item{\code{bc2000}}{double. Percent using contraception}
  \item{\code{bc2001}}{double. Percent using contraception}
  \item{\code{bc2002}}{double. Percent using contraception}
  \item{\code{bc2003}}{double. Percent using contraception}
  \item{\code{bc2004}}{double. Percent using contraception}
  \item{\code{bc2005}}{double. Percent using contraception}
  \item{\code{bc2006}}{double. Percent using contraception}
  \item{\code{bc2007}}{double. Percent using contraception}
  \item{\code{bc2008}}{double. Percent using contraception}
  \item{\code{bc2009}}{double. Percent using contraception}
  \item{\code{bc2010}}{double. Percent using contraception}
  \item{\code{bc2011}}{double. Percent using contraception}
  \item{\code{bc2012}}{double. Percent using contraception}
  \item{\code{bc2013}}{double. Percent using contraception}
  \item{\code{bc2014}}{double. Percent using contraception}
  \item{\code{bc2015}}{double. Percent using contraception}
  \item{\code{bc2016}}{double. Percent using contraception}
  \item{\code{bc2017}}{double. Percent using contraception}
  \item{\code{bc2018}}{double. Percent using contraception}
  \item{\code{bc2019}}{double. Percent using contraception}
  \item{\code{bc2020}}{double. Percent using contraception}
  \item{\code{bc2021}}{double. Percent using contraception}
  \item{\code{bc2022}}{double. Percent using contraception}
  \item{\code{bc2023}}{double. Percent using contraception}
  \item{\code{bc2024}}{double. Percent using contraception}
  \item{\code{bc2025}}{double. Percent using contraception}
  \item{\code{bc2026}}{double. Percent using contraception}
  \item{\code{bc2027}}{double. Percent using contraception}
  \item{\code{bc2028}}{double. Percent using contraception}
  \item{\code{bc2029}}{double. Percent using contraception}
  \item{\code{bc2030}}{double. Percent using contraception}
}
}
\source{
\url{http://www.un.org/en/development/desa/population/theme/family-planning/cp_model.shtml}

\url{http://dx.doi.org/10.1016/S0140-6736(12)62204-1}
}
\usage{
contraception
}
\description{
Percentage of married or in-union women aged 15 to 49 who are currently
using modern, traditional, or any method of contraception. Contraceptive prevalence
estimates and projections are based on the country-specific data compiled
in World Contraceptive Use 2015. Model results are as of March 2015.
}
\keyword{datasets}
