% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_centroids}
\alias{country_centroids}
\title{Country centroids (latitude and longitude)}
\format{
A data frame with 235 rows and 4 variables:
\describe{
  \item{\code{name}}{character. Country name}
  \item{\code{iso3}}{character. ISO Alpha 3}
  \item{\code{LAT}}{double. Latitude (decimal)}
  \item{\code{LONG}}{double. Longitude (decimal)}
}
}
\source{
\url{http://gothos.info/resource_files/country_centroids.zip}
}
\usage{
country_centroids
}
\description{
Latitude and longitude coordinates in decimal degrees for the
geographic center (centroids) for each nation in the world.
This includes all independent countries and most dependent
entities and territories. Most of the data for the world was
extracted from the NGA's GEOnet Names Server (GNS). Since the
United States and its territories are not included in this resource,
data for those areas was downloaded from the USGS Geographic Names
Information System (GNIS).
}
\examples{
# Distance matrix (in kilometers)
library(sp)
d <- spDists(as.matrix(country_centroids[c('LONG', 'LAT')]), longlat = T)
rownames(d) <- country_centroids$iso3
colnames(d) <- country_centroids$iso3
}
\keyword{datasets}
