% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{developing}
\alias{developing}
\title{Developing countries 2012}
\format{
A data frame with 187 rows and 6 variables:
\describe{
  \item{\code{region}}{character. Global Burden of Disease region.}
  \item{\code{gbd.super.region}}{character. Global Burden of Disease super-region}
  \item{\code{country}}{character. Country name}
  \item{\code{developing}}{logical. Is country developing (T) or developed (F).}
  \item{\code{iso3}}{character. ISO Alpha 3.}
  \item{\code{country.code}}{integer. ISO country code.}
}
}
\usage{
developing
}
\description{
Developing countries as defined in the GBD studies, along with regions and super-regions.
}
\keyword{datasets}
