% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gem}
\alias{gem}
\title{Gender Empowerment Measure 2009}
\format{
A data frame with 109 rows and 4 variables:
\describe{
  \item{\code{country}}{character. Country name.}
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{gem}}{double. Gender Empowerment Measure. Higher values indicate greater gender equality.}
}
}
\source{
\url{http://hdr.undp.org/sites/default/files/reports/269/hdr_2009_en_complete.pdf}
}
\usage{
gem
}
\description{
The Gender Empowerment Measure (GEM) is a composite indicator that captures
gender inequality in three key areas: the extent of women's political
participation and decision-making, economic participation and decision
making-power and the power exerted by women over economic resources.
}
\keyword{datasets}
