% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gggi}
\alias{gggi}
\title{Global Gender Gap Index 2014}
\format{
A data frame with 142 rows and 3 variables:
\describe{
  \item{\code{Country}}{character. Country name.}
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{GGGI}}{double. Global Gender Gap Index. Higher scores indicate more gender equality.}
}
}
\source{
\url{http://reports.weforum.org/global-gender-gap-report-2014/}
}
\usage{
gggi
}
\description{
Through the Global Gender Gap Report 2014, the World Economic Forum
quantifies the magnitude of gender-based disparities and tracks their
progress over time. While no single measure can capture the complete
situation, the Global Gender Gap Index presented in this Report seeks
to measure one important aspect of gender equality: the relative gaps
between women and men across four key areas: health, education, economy
and politics.
}
\keyword{datasets}
