% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infantmortality}
\alias{infantmortality}
\title{Infant mortality rate 1950-2015}
\format{
A data frame with 200 rows and 15 variables:
\describe{
  \item{\code{country}}{character. Country name}
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{im1950}}{double. Infant mortality 1950-1955}
  \item{\code{im1955}}{double. Infant mortality 1955-1960}
  \item{\code{im1960}}{double. Infant mortality 1960-1965}
  \item{\code{im1965}}{double. Infant mortality 1965-1970}
  \item{\code{im1970}}{double. Infant mortality 1970-1975}
  \item{\code{im1975}}{double. Infant mortality 1975-1980}
  \item{\code{im1980}}{double. Infant mortality 1980-1985}
  \item{\code{im1985}}{double. Infant mortality 1985-1990}
  \item{\code{im1990}}{double. Infant mortality 1990-1995}
  \item{\code{im1995}}{double. Infant mortality 1995-2000}
  \item{\code{im2000}}{double. Infant mortality 2000-2005}
  \item{\code{im2005}}{double. Infant mortality 2005-2010}
  \item{\code{im2010}}{double. Infant mortality 2010-2015}
}
}
\source{
\url{http://esa.un.org/unpd/wpp/Download/Standard/Mortality/}
}
\usage{
infantmortality
}
\description{
Infant deaths per 1000 live births. Probability of dying between birth and exact age 1.
Both sexes combined.
}
\keyword{datasets}
