% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso3}
\alias{iso3}
\title{ISO alpha 3 codes}
\format{
A data frame with 249 rows and 3 variables:
\describe{
  \item{\code{name}}{character. ISO country name}
  \item{\code{iso3}}{character. ISO Alpha 3 code.}
  \item{\code{country.code}}{integer. ISO country code.}
}
}
\source{
\url{https://github.com/lukes/ISO-3166-Countries-with-Regional-Codes/tree/master/slim-3}
}
\usage{
iso3
}
\description{
English name, numeric country code and alpha-3 code (e.g., NZL)
}
\keyword{datasets}
