% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Tobacco agriculture land use 1961-2013}
\format{
A data frame with 45026 rows and 8 variables:
\describe{
  \item{\code{iso3}}{character. Country code.}
  \item{\code{Item Code}}{integer. FAO code for item measured.}
  \item{\code{Item}}{character. Item measured}
  \item{\code{Year}}{integer. Year}
  \item{\code{Unit}}{character. Unit of measurement.}
  \item{\code{Value}}{double. Value of measurement}
  \item{\code{Flag}}{character. Data source}
  \item{\code{Flag Description}}{character. Data source.}
}
}
\source{
\url{http://www.fao.org/statistics/en/}
}
\usage{
land
}
\description{
Land use for tobacco production  from the Food and Agriculture Organization (FAO) of the United Nations.
}
\keyword{datasets}
