% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ng_smoking}
\alias{ng_smoking}
\title{National smoking prevalences by sex and year 1980-2012}
\format{
A data frame with 187 rows and 10 variables:
\describe{
  \item{\code{country}}{character. Country name}
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{f1980}}{double. Female smoking prevalence 1980}
  \item{\code{f1996}}{double. Female smoking prevalence 1996}
  \item{\code{f2006}}{double. Female smoking prevalence 2006}
  \item{\code{f2012}}{double. Female smoking prevalence 2012}
  \item{\code{m1980}}{double. Male smoking prevalence 1980}
  \item{\code{m1996}}{double. Male smoking prevalence 1996}
  \item{\code{m2006}}{double. Male smoking prevalence 2006}
  \item{\code{m2012}}{double. Male smoking prevalence 2012}
}
}
\source{
\url{http://dx.doi.org/10.1001/jama.2013.284692}
}
\usage{
ng_smoking
}
\description{
Smoking prevalence for 187 countries for 1980,
1996, 2006, and 2012. These are the data that are reported
in Ng et al. (2014) Smoking Prevalence and Cigarette Consumption in 187 countries, 1980-2012. JAMA.
}
\keyword{datasets}
