% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Total population by country and year}
\format{
A data frame with 264 rows and 60 variables:
\describe{
  \item{\code{Country Name}}{character. Country name.}
  \item{\code{Country Code}}{character. ISO 3 country code.}
  \item{\code{Indicator Name}}{character. World Bank indicator name.}
  \item{\code{Indicator Code}}{character. World Bank indicator code.}
  \item{\code{1960}}{double. Year.}
  \item{\code{1961}}{double. Year.}
  \item{\code{1962}}{double. Year.}
  \item{\code{1963}}{double. Year.}
  \item{\code{1964}}{double. Year.}
  \item{\code{1965}}{double. Year.}
  \item{\code{1966}}{double. Year.}
  \item{\code{1967}}{double. Year.}
  \item{\code{1968}}{double. Year.}
  \item{\code{1969}}{double. Year.}
  \item{\code{1970}}{double. Year.}
  \item{\code{1971}}{double. Year.}
  \item{\code{1972}}{double. Year.}
  \item{\code{1973}}{double. Year.}
  \item{\code{1974}}{double. Year.}
  \item{\code{1975}}{double. Year.}
  \item{\code{1976}}{double. Year.}
  \item{\code{1977}}{double. Year.}
  \item{\code{1978}}{double. Year.}
  \item{\code{1979}}{double. Year.}
  \item{\code{1980}}{double. Year.}
  \item{\code{1981}}{double. Year.}
  \item{\code{1982}}{double. Year.}
  \item{\code{1983}}{double. Year.}
  \item{\code{1984}}{double. Year.}
  \item{\code{1985}}{double. Year.}
  \item{\code{1986}}{double. Year.}
  \item{\code{1987}}{double. Year.}
  \item{\code{1988}}{double. Year.}
  \item{\code{1989}}{double. Year.}
  \item{\code{1990}}{double. Year.}
  \item{\code{1991}}{double. Year.}
  \item{\code{1992}}{double. Year.}
  \item{\code{1993}}{double. Year.}
  \item{\code{1994}}{double. Year.}
  \item{\code{1995}}{double. Year.}
  \item{\code{1996}}{double. Year.}
  \item{\code{1997}}{double. Year.}
  \item{\code{1998}}{double. Year.}
  \item{\code{1999}}{double. Year.}
  \item{\code{2000}}{double. Year.}
  \item{\code{2001}}{double. Year.}
  \item{\code{2002}}{double. Year.}
  \item{\code{2003}}{double. Year.}
  \item{\code{2004}}{double. Year.}
  \item{\code{2005}}{double. Year.}
  \item{\code{2006}}{double. Year.}
  \item{\code{2007}}{double. Year.}
  \item{\code{2008}}{double. Year.}
  \item{\code{2009}}{double. Year.}
  \item{\code{2010}}{double. Year.}
  \item{\code{2011}}{double. Year.}
  \item{\code{2012}}{double. Year.}
  \item{\code{2013}}{double. Year.}
  \item{\code{2014}}{double. Year.}
  \item{\code{2015}}{double. Year.}
}
}
\source{
\url{http://data.worldbank.org/indicator/SP.POP.TOTL}
}
\usage{
population
}
\description{
( 1 ) United Nations Population Division. World Population Prospects,
( 2 ) Census reports and other statistical publications from national statistical offices,
( 3 ) Eurostat: Demographic Statistics,
( 4 ) United Nations Statistical Division. Population and Vital Statistics Report ( various years ),
( 5 ) U.S. Census Bureau: International Database, and
( 6 ) Secretariat of the Pacific Community: Statistics and Demography Programme.
}
\keyword{datasets}
