% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production}
\alias{production}
\title{Tobacco production 1961-2013}
\format{
A data frame with 19813 rows and 8 variables:
\describe{
  \item{\code{iso3}}{integer. Country code}
  \item{\code{Element Code}}{integer. FAO codes for data types.}
  \item{\code{Element}}{character. "Area harvested", "Production", "Yield".}
  \item{\code{Year}}{integer. Year.}
  \item{\code{Unit}}{character. Unit of measurement.}
  \item{\code{Value}}{double. Value of measurement.}
  \item{\code{Flag}}{character. Data source.}
  \item{\code{Flag Description}}{character. Data source.}
}
}
\source{
\url{http://www.fao.org/statistics/en/}
}
\usage{
production
}
\description{
Tobacco production from the Food and Agriculture Organization (FAO) of the United Nations.
}
\details{
Measurements
* Area harvested: Data refer to the area from which a crop is gathered. Area harvested, therefore, excludes the area from which, although sown or planted, there was no harvest due to damage, failure, etc. It is usually net for temporary crops and some times gross for permanent crops. Net area differs from gross area insofar as the latter includes uncultivated patches, footpaths, ditches, headlands, shoulders, shelterbelts, etc.If the crop under consideration is harvested more than once during the year as a consequence of successive cropping (i.e. the same crop is sown or planted more than once in the same field during the year), the area is counted as many times as harvested. On the contrary, area harvested will be recorded only once in the case of successive gathering of the crop during the year from the same standing crops. With regard to mixed and associated crops, the area sown relating to each crop should be reported separately. When the mixture refers to particular crops, generally grains, it is recommended to treat the mixture as if it were a single crop; therefore, area sown is recorded only for the crop reported. Source: FAO Statistics Division
* Production: Figures relate to the total domestic production whether inside or outside the agricultural sector, i.e. it includes non-commercial production and production from kitchen gardens. Unless otherwise indicated, production is reported at the farm level for crop and livestock products (i.e. in the case of crops, excluding harvesting losses) and in terms of live weight for fish items (i.e. the actual ex-water weight at the time of the catch). All data shown relate to total meat production from both commercial and farm slaughter. Data are expressed in terms of dressed carcass weight, excluding offal and slaughter fats. Production of beef and buffalo meat includes veal; mutton and goat meat includes meat from lambs and kids; pig meat includes bacon and ham in fresh equivalent. Poultry meat includes meat from all domestic birds and refers, wherever possible, to ready-to-cook weight. Source: FAO Statistics Division
* Yield: Harvested production per unit of harvested area for crop products. In most of the cases yield data are not recorded but obtained by dividing the production data by the data on area harvested. Data on yields of permanent crops are not as reliable as those for temporary crops either because most of the area information may correspond to planted area, as for grapes, or because of the scarcity and unreliability of the area figures reported by the countries, as for example for cocoa and coffee. Source: FAO Statistics Division

Units
* Ha: hectares
* Hg/Ha: hectogrammes per hectare
* tonnes: metric tons
}
\keyword{datasets}
