% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking_age_sex}
\alias{smoking_age_sex}
\title{National smoking prevalences by age, sex, and year 1980-2013}
\format{
A data frame with 203456 rows and 7 variables:
\describe{
  \item{\code{iso3}}{character. ISO Alpha 3 code.}
  \item{\code{year}}{double. Year.}
  \item{\code{age_group}}{double. Age group.}
  \item{\code{sex}}{character. Sex}
  \item{\code{mean}}{double. Prevalence (fraction of the population, not percent).}
  \item{\code{lower}}{double. Lower 95\% CI}
  \item{\code{upper}}{double. Upper 95\% CI}
}
}
\source{
\url{http://dx.doi.org/10.1001/jama.2013.284692}
}
\usage{
smoking_age_sex
}
\description{
National smoking prevalences from 1980-2013, by 5-year age
groups from 15-99, and by sex. This file was obtained
from Murray's group via email request.
}
\details{
These are IHME results data from a global analysis of tobacco use
published in the Journal of the American Medical Association in January 2014.
The study, "Smoking Prevalence and Cigarette Consumption in 187 countries, 1980-2012,"
provides smoking prevalence, cigarette consumption, and annual rate of change
estimates by country and sex from 1980 to 2012.
}
\keyword{datasets}
