% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfr}
\alias{tfr}
\title{Total Fertility Rates 1950-2015}
\format{
A data frame with 200 rows and 16 variables:
\describe{
  \item{\code{country}}{character. Country name.}
  \item{\code{iso3}}{character. ISO Alpha 3.}
  \item{\code{tfr1950}}{double. TFR 1950-1955}
  \item{\code{tfr1955}}{double. TFR 1955-1960}
  \item{\code{tfr1960}}{double. TFR 1960-1965}
  \item{\code{tfr1965}}{double. TFR 1965-1970}
  \item{\code{tfr1970}}{double. TFR 1970-1975}
  \item{\code{tfr1975}}{double. TFR 1975-1980}
  \item{\code{tfr1980}}{double. TFR 1980-1985}
  \item{\code{tfr1985}}{double. TFR 1985-1990}
  \item{\code{tfr1990}}{double. TFR 1990-1995}
  \item{\code{tfr1995}}{double. TFR 1995-2000}
  \item{\code{tfr2000}}{double. TFR 2000-2005}
  \item{\code{tfr2005}}{double. TFR 2005-2010}
  \item{\code{tfr2010}}{double. TFR 2010-2015}
}
}
\source{
\url{http://esa.un.org/unpd/wpp/Download/Standard/Fertility/}
}
\usage{
tfr
}
\description{
Total fertility rates computed by the UN Department of Economic and Social
Affairs, Population Division. World Population Prospects. The 2015 Revision.
\url{http://esa.un.org/unpd/wpp/publications/files/key_findings_wpp_2015.pdf}.
Years are
}
\keyword{datasets}
