% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfrlong}
\alias{tfrlong}
\title{Total Fertility Rates 1950-2015 (long format)}
\format{
A data frame with 2400 rows and 3 variables:
\describe{
  \item{\code{iso3}}{character. ISO Alpha 3 country code.}
  \item{\code{year}}{double. Year.}
  \item{\code{tfr}}{double. Total fertility rate.}
}
}
\source{
\url{http://esa.un.org/unpd/wpp/Download/Standard/Fertility/}
}
\usage{
tfrlong
}
\description{
Total fertility rates computed by the UN Department of Economic and Social
Affairs, Population Division. World Population Prospects. The 2015 Revision.
\url{http://esa.un.org/unpd/wpp/publications/files/key_findings_wpp_2015.pdf}
}
\keyword{datasets}
