% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u5mortality}
\alias{u5mortality}
\title{Under 5 mortality rate 1950-2015}
\format{
A data frame with 200 rows and 15 variables:
\describe{
  \item{\code{country}}{character. Country name}
  \item{\code{iso3}}{character. ISO Alpha 3 code}
  \item{\code{u5m1950}}{double. Under 5 mortality 1950-1955}
  \item{\code{u5m1955}}{double. Under 5 mortality 1955-1960}
  \item{\code{u5m1960}}{double. Under 5 mortality 1960-1965}
  \item{\code{u5m1965}}{double. Under 5 mortality 1965-1970}
  \item{\code{u5m1970}}{double. Under 5 mortality 1970-1975}
  \item{\code{u5m1975}}{double. Under 5 mortality 1975-1980}
  \item{\code{u5m1980}}{double. Under 5 mortality 1980-1985}
  \item{\code{u5m1985}}{double. Under 5 mortality 1985-1990}
  \item{\code{u5m1990}}{double. Under 5 mortality 1990-1995}
  \item{\code{u5m1995}}{double. Under 5 mortality 1995-2000}
  \item{\code{u5m2000}}{double. Under 5 mortality 2000-2005}
  \item{\code{u5m2005}}{double. Under 5 mortality 2005-2010}
  \item{\code{u5m2010}}{double. Under 5 mortality 2010-2015}
}
}
\source{
\url{http://esa.un.org/unpd/wpp/Download/Standard/Mortality/}
}
\usage{
u5mortality
}
\description{
Child deaths per 1000 live births. Probability of dying between birth and exact age 5.
Both sexes combined.
}
\keyword{datasets}
